+++ smol/capp/generate/special/sqs.py
@@ -13,9 +13,9 @@ from smol._utils import derived_class_factory
 from smol.capp import enumerate_supercell_matrices, generate_random_ordered_occupancy
 from smol.cofe import ClusterSubspace
 from smol.moca import Ensemble, SampleContainer, Sampler
-from smol.moca._trace import Trace
 from smol.moca.kernel import MulticellMetropolis, mckernel_factory
 from smol.moca.processor.feature import FeatureDistanceProcessor
+from smol.moca.trace import Trace
 
 SQS = namedtuple("SQS", ["score", "features", "supercell_matrix", "structure"])