+++ smol/moca/kernel/mcusher.py
@@ -17,21 +17,21 @@ import numpy as np
 from monty.json import jsanitize
 from scipy.special import gammaln
 
-from smol._utils import class_name_from_str, derived_class_factory
-from smol.moca._metadata import Metadata
-from smol.moca._utils.math import (
+from smol.moca.composition import CompositionSpace
+from smol.moca.metadata import Metadata
+from smol.moca.utils.math import (
     NUM_TOL,
     choose_section_from_partition,
     flip_weights_mask,
 )
-from smol.moca._utils.occu import (
+from smol.moca.utils.occu import (
     delta_counts_from_step,
     get_dim_ids_by_sublattice,
     get_dim_ids_table,
     occu_to_counts,
     occu_to_species_list,
 )
-from smol.moca.composition import CompositionSpace
+from smol.utils.class_utils import class_name_from_str, derived_class_factory
 
 
 class MCUsher(ABC):