+++ smol/moca/sampler/sampler.py
@@ -12,11 +12,10 @@ from warnings import warn
 
 import numpy as np
 
-from smol._utils import progress_bar
-from smol.moca._trace import Trace
 from smol.moca.kernel import mckernel_factory
-from smol.moca.kernel._base import MulticellKernel
 from smol.moca.sampler.container import SampleContainer
+from smol.moca.trace import Trace
+from smol.utils.progressbar import progress_bar
 
 
 class Sampler: