+++ smol/utils/__init__.py
@@ -1,0 +1,4 @@
+"""General utilities and tools used in other modules.
+
+Functions and classes in this module are not meant only for internal use.
+"""